finishing touches

+++ conf.sh
@@ -1,30 +1,20 @@
 #!/bin/bash
 
-#read -p "Enter username: " username
-#read -s -p "Enter password: " password
-#echo
-#read -p "Enter your WiFi SSID: " wifi_ssid
-#read -s -p "Enter your WiFi password: " wifi_password
-#echo
-#read -p "Enter hostname: " hostname
-#read -p "Enter webpage to display on kiosk: " url
-#echo "Enter IP Address and Subnet-Mask in \"1.2.3.4/24\" format. "
-#read -p "Leave Empty for DHCP. IP: " ip_addr
+read -p "Enter username: " username
+read -s -p "Enter password: " password
+echo
+read -p "Enter your WiFi SSID: " wifi_ssid
+read -s -p "Enter your WiFi password: " wifi_password
+echo
+read -p "Enter hostname: " hostname
+read -p "Enter webpage to display on kiosk: " url
+echo "Enter IP Address and Subnet-Mask in \"1.2.3.4/24\" format. "
+read -p "Leave Empty for DHCP. IP: " ip_addr
 
-username="r3"
-password="r3alraum"
-wifi_ssid="realraum"
-wifi_password="r3alraum"
-hostname="touchme"
-url="homeassistant.realraum.at"
-ip_addr="192.168.127.248/24"
-ip_gw="192.168.127.254"
-ip_dns="192.168.127.254"
-
-#if [ -z "$username" ] || [ -z "$password" ] || [ -z "$wifi_ssid" ] || [ -z "$wifi_password" ] || [ -z "$hostname" ]; then
-#    echo "Invalid input. Exiting script."
-#    exit 1
-#fi
+if [ -z "$username" ] || [ -z "$password" ] || [ -z "$wifi_ssid" ] || [ -z "$wifi_password" ] || [ -z "$hostname" ]; then
+    echo "Invalid input. Exiting script."
+    exit 1
+fi
 
 cat <<EOF > 25-wlan.network
 [Match]
@@ -43,8 +33,8 @@ EOF
 echo "nameserver 1.1.1.1" > resolv.conf
 
 else
-#  read -p "Enter Gateway IP: " ip_gw
-#  read -p "Enter DNS Server IP: " ip_dns
+  read -p "Enter Gateway IP: " ip_gw
+  read -p "Enter DNS Server IP: " ip_dns
   
 cat <<EOF >> 25-wlan.network
 [Network]
@@ -71,14 +61,6 @@ network={
 EOF
 
 cat <<EOF > firstboot.sh
-
-read -p "Show IP config? (y/n)" showip
-
-if [ "\$showip" == "y" ]; then
-    ip addr
-fi
-
-read
 
 apt update && apt upgrade -y
 apt install xwayland cage sudo -y
